requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
blinker==1.6.2
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
db-dtypes==1.1.1
distlib==0.3.6
entrypoints==0.4
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
firebase==4.0.1
firebase-admin==6.1.0
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-bigquery==3.10.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gradio==3.28.1
gradio_client==0.1.4
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
httpcore==0.17.0
httplib2==0.22.0
httptools==0.5.0
httpx==0.24.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
nanoid==2.0.0
numpy==1.24.3
orjson==3.8.11
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.0
proto-plus==1.22.2
protobuf==4.22.4
psycopg2-binary==2.9.6
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
redis==4.5.4
requests==2.29.0
rsa==4.9
semantic-version==2.10.0
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.12
starkbank-ecdsa==2.2.0
starlette==0.26.1
tomli==2.0.1
toolz==0.12.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.22.0
virtualenv==20.21.0
watchfiles==0.19.0
websockets==11.0.2
Werkzeug==2.3.3
yarl==1.9.2
